projects: adrv9009: Add an example for using DAC DMA
This patch offers an example of how to send custom data using the DMA.
The custom data is a sine wave stored in the sine_lut_iq array. Each
element of the array is 32-bit word composed of 16 bits of I data and 16
bits of Q data. I and Q parts are 90 degrees of of phase.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>